AI telegram chatbot/AI bot/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.77.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "python-telegram-bot[job-queue]>=22.0",
    "telegram>=0.0.1",
]
